add warning system

+++ commands/Guild/warn.js
@@ -1,19 +1,26 @@
 module.exports.run = async (bot, message, args, settings) => {
   //get user for warning
-  let wUser = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]))
-  if (!wUser) {
-		message.delete()
-		message.channel.send("`Can't find user!`").then(m => m.delete({ timeout: 3500 }))
-		return
-	}
-  let wReason = (args.join(" ").slice(22)) ? args.join(" ").slice(22) : "No reason given"
+  if (message.deletable) message.delete()
   //Check to see if user has right permisisons
-	if (!message.member.hasPermission("KICK_MEMBERS") || wUser.hasPermission("KICK_MEMBERS")) {
-		message.delete()
-		message.channel.send("`You are unable to warn this person.`").then(m => m.delete({ timeout: 3500 }))
+	if (!message.member.hasPermission("KICK_MEMBERS")) {
+		message.channel.send({embed:{color:15158332, description:`${bot.config.emojis.cross} You are missing the permission: \`KICK_MEMBERS\`.`}}).then(m => m.delete({ timeout: 10000 }))
 		return
 	}
-  bot.warning(message, wUser, wReason, settings)
+  //get user to warn
+  let wUser = message.guild.member(message.mentions.users.first() || message.guild.members.cache.get(args[0]))
+  if (!wUser) {
+		message.channel.send({embed:{color:15158332, description:`${bot.config.emojis.cross} I was unable to find this user.`}}).then(m => m.delete({ timeout: 10000 }))
+		return
+	}
+  //make sure that the user that is getting warned has administrator permissions
+  if (wUser.hasPermission("ADMINISTRATOR")) {
+    message.channel.send({embed:{color:15158332, description:`${bot.config.emojis.cross} I am unable to warn this user.`}}).then(m => m.delete({ timeout: 10000 }))
+		return
+  }
+  //get reason for warning
+  let wReason = (args.join(" ").slice(22)) ? args.join(" ").slice(22) : "No reason given"
+  //Warning is sent to warning manager
+  require('../../modules/warning').run(bot, message, wUser, wReason, settings)
 }
 module.exports.config = {
 	command: "warn",
